Typescript stores

+++ src/App.tsx
@@ -27,7 +27,9 @@ function App() {
   ) => {
     setModalOpen(!modalOpen);
     if (id) setDeleteId(id);
-    if (replyingTo) setReplyTo(replyingTo);
+    if (replyingTo) {
+      setReplyTo(replyingTo);
+    } else setReplyTo("");
   };
   const comments = useSelector(selectComments);
 
@@ -38,9 +40,12 @@ function App() {
     }
   }, [dispatch, comments]);
 
-  const removeReplyHandler = () => dispatch(replyRemoved(comments, deleteId));
-  const removeCommentHandler = () =>
-    dispatch(commentRemoved(comments, deleteId));
+  const removeReplyHandler = () => {
+    if (deleteId) dispatch(replyRemoved(comments, deleteId));
+  };
+  const removeCommentHandler = () => {
+    if (deleteId) dispatch(commentRemoved(comments, deleteId));
+  };
 
   // resetStore();
   ///////////////////////////////////////////////
